refactor: move colmodernvbert related onnx embed to its class
